requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
